async/await and less logic per file
