+++ lib/apply-styles.js
@@ -1,0 +1,18 @@
+"use strict"
+
+module.exports = function (bundle, styles) {
+  styles.nodes = []
+
+  // Strip additional statements.
+  bundle.forEach(stmt => {
+    if (["charset", "import", "media"].includes(stmt.type)) {
+      stmt.node.parent = undefined
+      styles.append(stmt.node)
+    } else if (stmt.type === "nodes") {
+      stmt.nodes.forEach(node => {
+        node.parent = undefined
+        styles.append(node)
+      })
+    }
+  })
+}
